+++ vanilla-changes/vehicles-changes.lua
@@ -29,7 +29,15 @@ data:extend({{
   enabled = false,
   ingredients = {{"steel-plate", 15}, {"processing-unit", 5}, {"copper-cable", 20}},
   result = "space-train-battery-charging-station"
-}, {
+}, -- Experimental charging station
+{
+  type = "recipe",
+  name = "experimental-space-train-battery-charging-station",
+  energy_required = 10,
+  enabled = false,
+  ingredients = {{"steel-plate", 25}, {"processing-unit", 50}, {"copper-cable", 200}, {"space-train-battery-charging-station", 1}},
+  result = "experimental-space-train-battery-charging-station"
+},{
   type = "recipe",
   name = "space-train-battery-pack",
   category = "advanced-crafting",
@@ -44,7 +52,7 @@ data:extend({{
   name = "space-train-battery-pack-recharge",
   category = "electrical",
   hidden = true,
-  energy_required = 30,
+  energy_required = 60,
   enabled = false,
   icon = "__electric-trains__/graphics/icons/battery.png",
   icon_size = 128,
@@ -58,6 +66,30 @@ data:extend({{
   }, {
     name = "space-train-battery-pack",
     probability = 0.99,
+    amount = 1
+  }}
+
+},
+{
+  -- Faster space train battery park recharge
+  type = "recipe",
+  name = "faster-space-train-battery-pack-recharge",
+  category = "faster-electrical",
+  hidden = true,
+  energy_required = 20,
+  enabled = false,
+  icon = "__electric-trains__/graphics/icons/fast-battery.png",
+  icon_size = 128,
+  subgroup = "intermediate-product",
+  allow_as_intermediate = false,
+  ingredients = {{"space-train-discharged-battery-pack", 1}},
+  results = {{
+    name = "space-train-destroyed-battery-pack",
+    probability = 0.15,
+    amount = 1
+  }, {
+    name = "space-train-battery-pack",
+    probability = 0.85,
     amount = 1
   }}
 }})