+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-# Version: 0.1.1
-
-### 06.01.2024
+# Version: 0.1.1 - 06.01.2024
 
 - First version completed. (Forked from se-space-trains)
 - Rebranded research and added new infinite research.
 - Renamed all new entities and items.
 - New MiniMap icons for electric trains.
+
+# Version: 0.1.2 - 07.01.2024
+
+- Added Experimental Train Powerpack Charging Stations. (New research)
+- A new "Faster Train Powerpack Charging" recipe for the experimental charging station.
+
+### Some balance changes
+
+- All researches have sensible prerequisites and required science cost.
+- Experimental charging stations use a lot more power than regular charging stations.
+- Longer powerpack charge times.
+
+### Bug fixes